% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier_coords.R
\name{bezier_to_polygon}
\alias{bezier_to_polygon}
\title{Transform bezier dataframe to dataframe with path coordinates}
\usage{
bezier_to_polygon(df_benjamini_leaf, ...)
}
\arguments{
\item{df_benjamini_leaf}{Dataframe returned by \code{benjamini_leaf()}}

\item{...}{grouping variables in \code{df_benjamini_leaf} that will be kept in the transformation.}
}
\value{
Dataframe with the coordinates of the bezier curve interpolations.
}
\description{
Transform bezier dataframe to dataframe with path coordinates
}
\examples{
df_coords <- benjamini_leaf() \%>\%
  tidyr::unite(b, i_part, element, remove = FALSE) \%>\%
  bezier_to_polygon(b, i_part, element)
df_coords
df_coords \%>\%
  ggplot2::ggplot(ggplot2::aes(x = x, y = y, group = element, fill = element)) +
  ggplot2::geom_polygon()
}
