% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_bezier_skeleton}
\alias{df_bezier_skeleton}
\title{Example bezier dataframe to grow leaves on}
\format{
A data frame with 40 rows and 4 variables:
\describe{
\item{i_branch}{branch index}
\item{svg_point_type}{original point type in the svg file }
\item{x}{x coordinate}
\item{y}{y coordinate}
}
}
\usage{
df_bezier_skeleton
}
\description{
Generated with \code{vignette("import_svg_bezier")}
}
\keyword{datasets}
