% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gen_benjamini_points}
\alias{gen_benjamini_points}
\title{Title}
\usage{
gen_benjamini_points(
  x1 = 10,
  y1 = 40,
  dx21 = sample(12:20, 1),
  dy21 = sample(-4:-10, 1),
  dx32 = sample(10:18, 1),
  dy32 = runif(1, 0.88 * (-dy21 - 1), 0.92 * (-dy21 - 1)),
  dx43 = sample(4:6, 1),
  dy43 = y1 + dy21 + dy32
)
}
\arguments{
\item{dy43}{}
}
\value{

}
\description{
Title
}
