% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benjamini_leaves.R
\name{gen_benjamini_slopes}
\alias{gen_benjamini_slopes}
\title{Generate bezier slopes coordinates}
\usage{
gen_benjamini_slopes(leaf_params = gen_leaf_parameters())
}
\arguments{
\item{leaf_params}{parameters generated by \code{gen_leaf_parameters()}}
}
\value{
dataframe with sx1-4 in the x variable and sy1-4 in the y variable
}
\description{
Generate bezier slopes coordinates
}
\examples{
gen_benjamini_slopes()
}
