% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier_coords.R
\name{gen_leaf_bezier_coords}
\alias{gen_leaf_bezier_coords}
\title{Transform bezier dataframe to dataframe with path coordinates}
\usage{
gen_leaf_bezier_coords(df_benjamini_leaf)
}
\arguments{
\item{df_benjamini_leaf}{Dataframe returned by \code{benjamini_leaf()}}
}
\value{
Dataframe with the coordinates of the bezier curve interpolations.
}
\description{
Transform bezier dataframe to dataframe with path coordinates
}
\examples{
df_coords <- benjamini_leaf() \%>\%
  gen_leaf_bezier_coords()
df_coords
df_coords \%>\%
  ggplot2::ggplot(ggplot2::aes(x= x, y = y, fill = half)) +
  ggplot2::geom_polygon()
}
