% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benjamini_leaves.R
\name{gen_leaf_parameters}
\alias{gen_leaf_parameters}
\title{Generate bezier coordinates of a leaf}
\usage{
gen_leaf_parameters(
  x1 = 10,
  y1 = 40,
  dx21 = sample(12:20, 1),
  dy21 = sample(-4:-10, 1),
  dx32 = sample(10:18, 1),
  dy32 = stats::runif(1, 0.92 * (-dy21 - 1), 0.95 * (-dy21 - 1)),
  dx43 = sample(4:6, 1),
  dy43 = y1 + dy21 + dy32,
  sx1 = sample(1:3, 1),
  sx2 = sample(4:6, 1),
  sx3 = sample(2:4, 1),
  sx4 = stats::runif(1, 0, 0.2),
  sy1 = sample(-4:-6, 1),
  sy2 = stats::runif(1, -0.5, 0.5),
  sy3 = stats::runif(1, 0.5, 1.5),
  sy4 = stats::runif(1, 0.2, 0.7),
  smx1 = sample(-5:-15, 1),
  smx2 = sample(-5:-15, 1),
  smy1 = stats::runif(1, -1, 1),
  smy2 = stats::runif(1, -1, 1)
)
}
\arguments{
\item{x1, y1}{coordinates of the leaf origin}

\item{dx21, dy21, dx32, dy32, dx43, dy43}{coordinates of the other control points (relative to the origin)}

\item{sx1, sx2, sx3, sx4}{x coordinates of the control points}

\item{sy1, sy2, sy3, sy4}{y coordinates of the control points}

\item{smx1, smx2, smy1, smy2}{x & y coordinates of the middle vein control points}
}
\value{
named list with all parameters
}
\description{
Generate bezier coordinates of a leaf
}
\examples{
gen_leaf_parameters()
}
