% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benjamini_leaves.R
\name{gen_middle_line_slopes}
\alias{gen_middle_line_slopes}
\title{Generate bezier slopes of the line in the middle of the leaf}
\usage{
gen_middle_line_slopes(
  sx1 = sample(-5:-15, 1),
  sx2 = sample(-5:-15, 1),
  sy1 = stats::runif(1, -1, 1),
  sy2 = stats::runif(1, -1, 1)
)
}
\arguments{
\item{sx1, sx2, sy1, sy2}{x & y coordinates of the middle line bezier curve of the benjamini leaf}
}
\value{

}
\description{
Generate bezier slopes of the line in the middle of the leaf
}
\examples{
gen_middle_line_slopes()
}
