% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{get_one_bezier}
\alias{get_one_bezier}
\title{Title}
\usage{
get_one_bezier(i, points_df, slopes_df)
}
\arguments{
\item{slopes_df}{}
}
\value{

}
\description{
Title
}
\examples{
set.seed(123)
points_df <- gen_benjamini_points()
slopes_df <- gen_benjamini_slopes()
get_one_bezier(1, points_df, slopes_df)
}
