% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R
\name{spark_norm}
\alias{spark_norm}
\title{Manipulate the angles of the leaves with a normal distribution}
\usage{
spark_norm(mean = 0, sd = 3)
}
\arguments{
\item{mean, sd}{Parameters passed to \code{stats::dnorm()}.}
}
\value{
dnorm() function with n_leaves as one of the arguments
}
\description{
This function returns a function which itself returns a numerical vector of
length of the number of leaves on the branch.
}
\examples{
spark_norm()
}
