% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R
\name{spark_weibull}
\alias{spark_weibull}
\title{Manipulate the sizes of the leaves with a Weibull distribution}
\usage{
spark_weibull(shape = 1.2, scale_factor = 0.5)
}
\arguments{
\item{shape, scale_factor}{Parameters passed to \code{stats::dweibull()}.}
}
\value{
dweibull() function with n_leaves as one of the arguments
}
\description{
This function returns a function which itself returns a numerical vector of
length of the number of leaves on the branch.
}
\examples{
spark_weibull()
}
