% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R
\name{benjamini_branch}
\alias{benjamini_branch}
\title{Generate a branch of benjamini leaves}
\usage{
benjamini_branch(
  df_bezier = tibble::tibble(x = c(70, 84, 126, 168), y = c(280, 245, 217, 217)),
  leaf_dist_approx = 20,
  leaf_angle = 45,
  first_dir = sample(0:1, 1),
  stalk_len = 15
)
}
\arguments{
\item{leaf_dist_approx}{approximate distance of two leaves}

\item{leaf_angle}{angle between the leaf stalks and the branch}

\item{first_dir}{direction of the first leaf in the branch (0 for right; 1 for left)}

\item{stalk_len}{length of a stalk}

\item{df_branch}{dataframe containing the x & y coordinates of a branch}
}
\value{
A dataframe conaining the data for leaves on a branch (see example).
}
\description{
Generate a branch of benjamini leaves
}
\examples{
benjamini_branch() \%>\%
  ggplot2::ggplot(ggplot2::aes(x = x1, y = y1, xend = x2, yend = y2)) +
  ggplot2::geom_segment() +
  ggforce::geom_bezier(ggplot2::aes(x = x, y = y, group = i)) +
  ggplot2::coord_equal()
}
