% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benjamini_leaves.R
\name{benjamini_leaf}
\alias{benjamini_leaf}
\title{Generate bezier curve coordinates of a benjamini leaf}
\usage{
benjamini_leaf(
  leaf_params = gen_leaf_parameters(),
  omega = 0,
  xrot = gen_benjamini_points(leaf_params)$x[1],
  yrot = gen_benjamini_points(leaf_params)$y[1],
  precision = 2
)
}
\arguments{
\item{leaf_params}{parameter that control the leaf shape}

\item{omega}{rotation angle of the leaf}

\item{xrot}{rotation x coordinate (preset to leaf origin).}

\item{yrot}{rotation y coordinate (preset to leaf origin).}

\item{precision}{numeric precision of the output}
}
\value{
A dataframe conaining the data for the bezier curves of a leaf (see example).
}
\description{
Generate bezier curve coordinates of a benjamini leaf
}
\examples{
df <- benjamini_leaf()
df
df \%>\%
  # This generated a unique identifier for the 4 rows of each bezier curve:
  tidyr::unite(b, element, i_part, remove = FALSE) \%>\%
  ggplot2::ggplot() +
  ggforce::geom_bezier(ggplot2::aes(x = x, y = y, group = b))
}
