% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benjamini_leaves.R
\name{benjamini_leaf}
\alias{benjamini_leaf}
\title{Generate bezier curve coordinates of a benjamini leaf}
\usage{
benjamini_leaf(
  points_df = gen_benjamini_points(),
  slopes_df = gen_benjamini_slopes(),
  slopes_middle_df = gen_middle_line_slopes(),
  omega = 0,
  xrot = points_df$x[1],
  yrot = points_df$y[1],
  precision = 2
)
}
\arguments{
\item{points_df}{dataframe generated by \code{gen_benjamini_points()} (see
example).}

\item{slopes_df}{dataframe generated by \code{gen_benjamini_slopes()} (see
example).}

\item{slopes_middle_df}{slopes of the bezier curve separating the leaf in the middle}

\item{omega}{rotation angle of the leaf}

\item{xrot}{rotation x coordinate}

\item{yrot}{rotation y coordinate}

\item{precision}{numeric precision of the output}
}
\value{
A dataframe conaining the data for the bezier curves of a leaf (see example).
}
\description{
Generate bezier curve coordinates of a benjamini leaf
}
\examples{
df <- benjamini_leaf()
df
ggplot2::ggplot(df) +
  ggforce::geom_bezier(ggplot2::aes(x = x, y = y, group = i))
}
