% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gen_benjamini_points}
\alias{gen_benjamini_points}
\title{Generate bezier end points}
\usage{
gen_benjamini_points(
  x1 = 10,
  y1 = 40,
  dx21 = sample(12:20, 1),
  dy21 = sample(-4:-10, 1),
  dx32 = sample(10:18, 1),
  dy32 = stats::runif(1, 0.92 * (-dy21 - 1), 0.95 * (-dy21 - 1)),
  dx43 = sample(4:6, 1),
  dy43 = y1 + dy21 + dy32
)
}
\arguments{
\item{x1, y1}{coordinates of the leaf origin}

\item{dx21, dy21, dx32, dy32, dx43, dy43}{coordinates of the other control points (relative to the origin)}
}
\value{
tibble with absolute coordinates
}
\description{
Generate bezier end points
}
\examples{
gen_benjamini_points()
}
