% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benjamini_leaves.R
\name{gen_benjamini_points}
\alias{gen_benjamini_points}
\title{Generate bezier end points}
\usage{
gen_benjamini_points(leaf_params = gen_leaf_parameters())
}
\arguments{
\item{leaf_params}{parameters generated by \code{gen_leaf_parameters()}}
}
\value{
tibble with absolute coordinates
}
\description{
Generate bezier end points
}
\examples{
gen_benjamini_points()
}
