% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gen_benjamini_slopes}
\alias{gen_benjamini_slopes}
\title{Title}
\usage{
gen_benjamini_slopes(
  sx1 = sample(1:3, 1),
  sx2 = sample(4:6, 1),
  sx3 = sample(2:4, 1),
  sx4 = runif(1, 0, 0.2),
  sy1 = sample(-4:-6, 1),
  sy2 = runif(1, -0.5, 0.5),
  sy3 = runif(1, 0.5, 1.5),
  sy4 = runif(1, 0.5, 1.5)
)
}
\arguments{
\item{sy4}{}
}
\value{

}
\description{
Title
}
