% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benjamini_leaves.R
\name{gen_middle_line_slopes}
\alias{gen_middle_line_slopes}
\title{Generate bezier slopes of the line in the middle of the leaf}
\usage{
gen_middle_line_slopes(leaf_params = gen_leaf_parameters())
}
\arguments{
\item{leaf_params}{parameters generated by \code{gen_leaf_parameters()}}
}
\value{
A dataframe containing the coordinates of the two control points of
the bezier curve defining the midvein of the leaf.
}
\description{
Generate bezier slopes of the line in the middle of the leaf
}
\examples{
gen_middle_line_slopes()
}
