% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{get_one_bezier}
\alias{get_one_bezier}
\title{Generate a dataframe of one bezier curve}
\usage{
get_one_bezier(i, points_df, slopes_df)
}
\arguments{
\item{i}{number of the bezier}

\item{points_df}{dataframe generated by \code{gen_benjamini_points()} (see
example).}

\item{slopes_df}{dataframe generated by \code{gen_benjamini_slopes()} (see
example).}
}
\value{

}
\description{
Generate a dataframe of one bezier curve
}
\examples{
set.seed(123)
points_df <- gen_benjamini_points()
slopes_df <- gen_benjamini_slopes()
df_bezier <- get_one_bezier(1, points_df, slopes_df)
ggplot2::ggplot(df_bezier, ggplot2::aes(x = x, y = y)) + ggforce::geom_bezier()
}
